+++ database/migrations/03_roles.ts
@@ -4,15 +4,18 @@ export default class extends BaseSchema {
   protected tableName = 'roles'
 
   public async up() {
-    this.schema.createTable(this.tableName, (table) => {
-      table.increments('id') //Primary key
-      table.string('name', 255).notNullable() //Rolename
-      table.timestamps(true) //timestamps (created_at and updated_at)
+    if (!this.schema.hasTable(this.tableName)) {
+      this.schema.createTable(this.tableName, (table) => {
+        table.increments('id') //Primary key
+        table.string('name', 255).notNullable() //Rolename
+        table.timestamps(true) //timestamps (created_at and updated_at)
 
-    })
+      })
+    }
   }
-
- public  async down() {
-    this.schema.dropTable(this.tableName)
-  }
-}''
+  public async down() {
+    if (await this.schema.hasTable(this.tableName)) {
+      this.schema.dropTable(this.tableName)
+    }
+  } 
+}
